Rename symbols to compliant with CamelCase requirement
